src/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django-environ==0.4.5
flake8==3.8.4
gunicorn==20.0.4
requests==2.25.1
psycopg2-binary==2.8.6
ptvsd==4.3.2 # Required for debugging.
